requirements.txt
requests>=2.31.0
loguru>=0.7.0
pandas>=2.0.0
streamlit>=1.35.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
